// src/app/shared/models/eveningObservation.model.ts
export class EveningObservation {
  presentWt?: string;
  wtResult?: string;
  bsResult?: string;
  bpType: string;
  presentBSreading?: string;
  symptomsBS?: string;
  BSzoneofpat?: string;
  BScreatedAt?: string;
  symptomsWt?: string;
  permittedWeightlwr?: string;
  permittedWeighthgr?: string;
  timepostdialysis?: string;
  presentBp?: string;
  bpResult?: string;
  symptomsBp?: string;
  status?: string;
  zoneOfPat?: string;
  createdAt?: string;
  weightZoneofPatient?: string;
  BPZoneofPatient?: string;
  weightCreatedAt?: string;
  BPCreatedAt?: string;
  indicators?: string;
}