// src/app/shared/models/morningObservation.model.ts
export class MorningObservation {
    presentWt?: string;
    wtResult: string;
    symptomsWt: string;
    bpType:string;
    bsResult?:string
    presentBSreading?:string
    BScreatedAt?:string
    symptomsBS?:string
    BSzoneofpat?:string
    permittedWeightlwr: string;
    permittedWeighthgr: string;
    timepostdialysis?: string;
    presentBp?: string;
    bpResult?: string;
    symptomsBp?: string;
    status?: string;
    zoneOfPatient?: string;
    createdAt?: string;
    weightZoneofPatient?: string;
    BPZoneofPatient?: string;
    weightCreatedAt?: string;
    BPCreatedAt?: string;
    indicators?:string;

}